requirements.txt
autopep8==1.5.4
chromedriver-binary==87.0.4280.88.0
docopt==0.6.2
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==2.4.11
pycodestyle==2.6.0
python-dotenv==0.15.0
records==0.5.3
selenium==3.141.0
SQLAlchemy==1.3.22
tablib==3.0.0
toml==0.10.2
urllib3==1.26.2
